+++ src/demo_uniswap/__init__.py
@@ -1,3 +1,0 @@
-import logging
-
-logging.getLogger(__name__).setLevel(logging.DEBUG)